temp_egonet_requirements.txt
flask
requests
python-dotenv
atproto
networkx
gunicorn
